requirements.txt
Django==4.2.1
django-rest-framework
django-cors-headers
# mysqlclient
psycopg2-binary